server/requirements.txt
PyYAML==6.0.1
websockets==9.1
